Sprint 20 - 2й commit. Много рефакторинга - разнес LocalStorage на несколько стораджей, соответственно разнес интеракторы и репозитории для истории, избранного и поиска...
