+++ app/src/main/java/com/example/playlistmaker/App.kt
@@ -5,6 +5,7 @@ import androidx.appcompat.app.AppCompatDelegate
 import com.example.playlistmaker.di.UiModule
 import com.example.playlistmaker.di.dataModule
 import com.example.playlistmaker.di.domainModule
+import com.example.playlistmaker.di.repositoryModule
 import com.example.playlistmaker.di.viewModelModule
 import com.example.playlistmaker.domain.settings.SettingsInteractor
 import org.koin.android.ext.android.inject
@@ -14,22 +15,21 @@ import org.koin.core.context.startKoin
 
 class App : Application() {
 
-    private val settingsInteractor: SettingsInteractor by inject()
-
     override fun onCreate() {
         super.onCreate()
 
         startKoin {
             androidContext(this@App)
-            modules(dataModule, domainModule, viewModelModule, UiModule)
+            modules(dataModule, repositoryModule, domainModule, viewModelModule, UiModule)
         }
+
+        val settingsInteractor: SettingsInteractor by inject()
 
         setDarkTheme(settingsInteractor.getThemeSettings().isDarkTheme)
     }
 
     companion object {
         const val LOCAL_STORAGE = "LOCAL_STORAGE"
-        const val CURRENT_TRACK = "current_track"
 
         fun setDarkTheme(darkThemeEnabled: Boolean) {
