+++ app/src/main/java/com/example/playlistmaker/data/FavoritesStorage.kt
@@ -1,0 +1,13 @@
+package com.example.playlistmaker.data
+
+interface FavoritesStorage {
+
+    fun addToFavorites(trackId: Int)
+
+    fun removeFromFavorites(trackId: Int)
+
+    fun changeFavorites(trackId: Int, remove: Boolean)
+
+    fun getSavedFavorites(): Set<String>
+
+}
