+++ app/src/main/java/com/example/playlistmaker/data/search/HistoryStorage.kt
@@ -1,0 +1,14 @@
+package com.example.playlistmaker.data.search
+
+import com.example.playlistmaker.data.dto.Response
+import com.example.playlistmaker.data.dto.TrackDto
+
+interface HistoryStorage {
+
+    fun getTracksFromHistory(): Response
+
+    fun addTrackToHistory(track: TrackDto)
+
+    fun clearTracksHistory()
+
+}
