+++ app/src/main/java/com/example/playlistmaker/data/settings/SettingsStorage.kt
@@ -1,10 +1,11 @@
-package com.example.playlistmaker.domain.settings
+package com.example.playlistmaker.data.settings
 
 import com.example.playlistmaker.domain.settings.model.ThemeSettings
 
-interface SettingsRepository {
+interface SettingsStorage {
 
     fun getThemeSettings(): ThemeSettings
+
     fun updateThemeSetting(settings: ThemeSettings)
 
 }
