+++ app/src/main/java/com/example/playlistmaker/di/DomainModule.kt
@@ -1,52 +1,39 @@
 package com.example.playlistmaker.di
 
-import com.example.playlistmaker.data.player.impl.PlayerRepositoryImpl
-import com.example.playlistmaker.data.search.impl.SearchRepositoryImpl
-import com.example.playlistmaker.data.settings.impl.SettingsRepositoryImpl
-import com.example.playlistmaker.data.sharing.impl.ExternalNavigatorImpl
 import com.example.playlistmaker.domain.player.PlayerInteractor
-import com.example.playlistmaker.domain.player.PlayerRepository
 import com.example.playlistmaker.domain.player.impl.PlayerInteractorImpl
+import com.example.playlistmaker.domain.search.FavoritesInteractor
+import com.example.playlistmaker.domain.search.HistoryInteractor
 import com.example.playlistmaker.domain.search.SearchInteractor
-import com.example.playlistmaker.domain.search.SearchRepository
+import com.example.playlistmaker.domain.search.impl.FavoritesInteractorImpl
+import com.example.playlistmaker.domain.search.impl.HistoryInteractorImpl
 import com.example.playlistmaker.domain.search.impl.SearchInteractorImpl
 import com.example.playlistmaker.domain.settings.SettingsInteractor
-import com.example.playlistmaker.domain.settings.SettingsRepository
 import com.example.playlistmaker.domain.settings.impl.SettingsInteractorImpl
-import com.example.playlistmaker.domain.sharing.ExternalNavigator
 import com.example.playlistmaker.domain.sharing.SharingInteractor
 import com.example.playlistmaker.domain.sharing.impl.SharingInteractorImpl
-import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
 val domainModule = module {
 
-    single<SettingsRepository> {
-        SettingsRepositoryImpl(get())
-    }
-
     single<SettingsInteractor> {
         SettingsInteractorImpl(get())
-    }
-
-    single<ExternalNavigator> {
-        ExternalNavigatorImpl(androidContext())
     }
 
     single<SharingInteractor> {
         SharingInteractorImpl(get())
     }
 
-    single<SearchRepository> {
-        SearchRepositoryImpl(get(), get())
-    }
-
     single<SearchInteractor> {
         SearchInteractorImpl(get())
     }
 
-    single<PlayerRepository> {
-        PlayerRepositoryImpl()
+    single<HistoryInteractor> {
+        HistoryInteractorImpl(get())
+    }
+
+    single<FavoritesInteractor> {
+        FavoritesInteractorImpl(get())
     }
 
     single<PlayerInteractor> {
