+++ app/src/main/java/com/example/playlistmaker/di/UiModule.kt
@@ -2,12 +2,13 @@ package com.example.playlistmaker.di
 
 import com.example.playlistmaker.util.ResourceProvider
 import com.example.playlistmaker.util.ResourceProviderImpl
+import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
 val UiModule = module {
 
     single<ResourceProvider> {
-        ResourceProviderImpl(get())
+        ResourceProviderImpl(androidContext())
     }
 
 }
