+++ app/src/main/java/com/example/playlistmaker/domain/search/FavoritesInteractor.kt
@@ -1,0 +1,13 @@
+package com.example.playlistmaker.domain.search
+
+interface FavoritesInteractor {
+
+    fun addToFavorites(trackId: Int)
+
+    fun removeFromFavorites(trackId: Int)
+
+    fun changeFavorites(trackId: Int, remove: Boolean)
+
+    fun getSavedFavorites(): Set<String>
+
+}
