+++ app/src/main/java/com/example/playlistmaker/domain/search/FavoritesRepository.kt
@@ -1,0 +1,13 @@
+package com.example.playlistmaker.domain.search
+
+interface FavoritesRepository {
+
+    fun addToFavorites(trackId: Int)
+
+    fun removeFromFavorites(trackId: Int)
+
+    fun changeFavorites(trackId: Int, remove: Boolean)
+
+    fun getSavedFavorites(): Set<String>
+
+}
