+++ app/src/main/java/com/example/playlistmaker/domain/search/HistoryInteractor.kt
@@ -1,0 +1,17 @@
+package com.example.playlistmaker.domain.search
+
+import com.example.playlistmaker.domain.models.Track
+
+interface HistoryInteractor {
+
+    fun addTrackToHistory(track: Track)
+
+    fun clearTracksHistory()
+
+    fun getTracksFromHistory(consumer: TracksConsumer)
+
+    interface TracksConsumer {
+        fun consume(foundTracks: List<Track>?, errorCode: Int?)
+    }
+
+}
