+++ app/src/main/java/com/example/playlistmaker/domain/search/SearchInteractor.kt
@@ -7,20 +7,4 @@ interface SearchInteractor {
 
     fun searchTracks(expression: String): Flow<Pair<List<Track>?, Int?>>
 
-    fun addTrackToFavorites(track: Track)
-
-    fun removeTrackFromFavorites(track: Track)
-
-    fun addTrackToHistory(track: Track)
-
-    fun removeTrackFromHistory(track: Track)
-
-    fun clearTracksHistory()
-
-    fun getTracksFromHistory(consumer: TracksConsumer)
-
-    interface TracksConsumer {
-        fun consume(foundTracks: List<Track>?, errorCode: Int?)
-    }
-
 }
