+++ app/src/main/java/com/example/playlistmaker/domain/search/impl/SearchInteractorImpl.kt
@@ -6,11 +6,8 @@ import com.example.playlistmaker.domain.search.SearchRepository
 import com.example.playlistmaker.util.Resource
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import java.util.concurrent.Executors
 
 class SearchInteractorImpl(private val repository: SearchRepository) : SearchInteractor {
-
-    private val executor = Executors.newCachedThreadPool()
 
     override fun searchTracks(expression: String): Flow<Pair<List<Track>?, Int?>> {
 
@@ -28,37 +25,4 @@ class SearchInteractorImpl(private val repository: SearchRepository) : SearchInt
         }
     }
 
-    override fun addTrackToFavorites(track: Track) {
-        repository.addTrackToFavorites(track)
-    }
-
-    override fun removeTrackFromFavorites(track: Track) {
-        repository.removeTrackFromFavorites(track)
-    }
-
-    override fun addTrackToHistory(track: Track) {
-        repository.addTrackToHistory(track)
-    }
-
-    override fun removeTrackFromHistory(track: Track) {
-        repository.removeTrackFromHistory(track)
-    }
-
-    override fun clearTracksHistory() {
-        repository.clearTracksHistory()
-    }
-
-    override fun getTracksFromHistory(consumer: SearchInteractor.TracksConsumer) {
-        executor.execute {
-            when (val resource = repository.getTracksFromHistory()) {
-                is Resource.Success -> {
-                    consumer.consume(resource.data, null)
-                }
-
-                is Resource.Error -> {
-                    consumer.consume(null, resource.code)
-                }
-            }
-        }
-    }
 }
