+++ app/src/main/java/com/example/playlistmaker/domain/settings/SettingsInteractor.kt
@@ -3,6 +3,7 @@ package com.example.playlistmaker.domain.settings
 import com.example.playlistmaker.domain.settings.model.ThemeSettings
 
 interface SettingsInteractor {
+
     fun getThemeSettings(): ThemeSettings
     fun updateThemeSetting(settings: ThemeSettings)
