+++ app/src/main/java/com/example/playlistmaker/domain/settings/impl/SettingsInteractorImpl.kt
@@ -5,6 +5,7 @@ import com.example.playlistmaker.domain.settings.SettingsRepository
 import com.example.playlistmaker.domain.settings.model.ThemeSettings
 
 class SettingsInteractorImpl(private val repository: SettingsRepository) : SettingsInteractor {
+
     override fun getThemeSettings(): ThemeSettings {
         return repository.getThemeSettings()
     }
